API/requirements.txt
Flask
Flask-Cors
python-dotenv
face_recognition
numpy
Pillow